MongoDB Atlas working
You need to install dnspython with pip to make it work with mongodb.
To enter MongoDB Atlas the email is developmentMastersPacman@gmail.com (Enter with Google) and the password is pacman2020*
In pipelines.py we also specify that the database is in a MongoServer Atlas in which the user of the database is "user" and the password is "pass". The database name is tiktok.
In the file pipelines.py it checks if the user is already inserted in the database  and if it's not, it inserts the user.